[brian] wiremock + basic tests

+++ app/domains/Properties.scala
@@ -3,7 +3,7 @@ package domains
 import play.api.libs.json.Json._
 
 case class Properties( country: String,
-                       result_count: Int,
+                       result_count: BigDecimal,
                        area_name: String,
                        listing: Seq[Property],
                        street: String,
